Machine base class
This PR adds a machine base class that gives machines an inventory items can be added to and removed from. It includes functions for mapping the index of an item in the array to a key that can then be matched with a recipe to see if it has enough to craft it, a function for getting the item name and index, and a function to craft the item. It also sets up a workflow for using data tables to instances items and recipes from